+++ src/singleTable/adaptor/definitions/index.ts
@@ -1,3 +1,4 @@
 export * from './operations';
 export * from './config';
 export * from './key';
+export * from './tableIndex/numericRef';