+++ src/singleTable/adaptor/definitions/tableIndex/index.ts
@@ -1,1 +1,2 @@
 export * from './tableIndex';
+export * from './numericRef';